move prop related functions into compile-props
